Alpha-Build

+++ app/controllers/twt_print.py
@@ -1,10 +1,12 @@
+import random
+import time
 import tweepy
 
 
-ckey= ''
-csecret= ''
-atoken= ''
-asecret= ''
+ckey= '*'
+csecret= '*'
+atoken= '*'
+asecret= '*'
 
 class TwtPrinter:
     def printTweet(self, user, text):
@@ -12,13 +14,15 @@ class TwtPrinter:
         auth = tweepy.OAuthHandler(ckey, csecret)
         auth.set_access_token(atoken, asecret)
         api = tweepy.API(auth)
+        message = "@" + user  + " " + text
         for status in tweepy.Cursor(api.user_timeline).items():
-            try:
-                api.destroy_status(status.id)
-            except:
-                pass
+            if status.text in message:
+                try:
+                    api.destroy_status(status.id)
+                except:
+                    pass
         # message = "@" + user  + " " + text
-        message = user  + " " + text
+        time.sleep(random.randint(1,3))
         print(message)
         api.update_status(status=message)
         # return
